moved calibration to its own file
